% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\docType{data}
\name{cities}
\alias{cities}
\title{City data}
\format{list of lists}
\usage{
cities
}
\description{
City data for the haterz project. Inlcludes data pertaining
    to city county names and coordinates.
}
\examples{
  cities
}
\keyword{datasets}
