% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{diversity_index}
\alias{diversity_index}
\title{Calculates the diversity index of a vector of data.}
\usage{
diversity_index(x, index = "shannon")
}
\arguments{
\item{x}{Vector of counts}

\item{index}{Selected diversity index. Can be 'shannon' or 'simpson'}
}
\value{
An index of class numeric
}
\description{
Calculates the diversity index of a vector of data.
}
\examples{
diversity_index(x = c(1, 2, 3))
diversity_index(x = c(20,20,20), index = 'simpson')
}
