% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{equitability}
\alias{equitability}
\title{Calculates the eveness of the diversity index. It can take on a value between
    0 and 1 with 1 being complete eveness.}
\usage{
equitability(x, index = "shannon")
}
\arguments{
\item{x}{Vector of counts}

\item{index}{Selected diversity index. Can be 'shannon' or 'simpson'}
}
\value{
A eveness metric of class numeric
}
\description{
Calculates the eveness of the diversity index. It can take on a value between
    0 and 1 with 1 being complete eveness.
}
\examples{
equitability(x = c(1, 2, 3))
equitability(x = c(20,20,20), index = 'simpson')
}
