% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper_functions.R
\name{gather_counties}
\alias{gather_counties}
\title{Gather ACS attributes from multiple counties}
\usage{
gather_counties(counties, variables, geography = "tract", year = 2015,
  geometry = FALSE)
}
\arguments{
\item{counties}{A vector of counties to collect data on. Must be in the following format
`SA,County Name` where SA is the state abbreviation and the Count Name is the name of
the county. For example, `MO,Jackson` would collect on Jackson county, Missouri.}

\item{variables}{A list of ACM coded variables to collect.}

\item{geography}{The geographical unit to collect on.}

\item{year}{The ACS collection year.}

\item{geometry}{An option to include the simple feature for mapping functionality.}
}
\value{
a simple feature data frame.
}
\description{
Gather ACS attributes from multiple counties
}
\examples{
gather_counties(counties = c('MO,Clay', 'MO,Jackson'),
                variables = c('B02001_001E', 'B02001_003E'))
gather_counties('MO,Jackson', variables = 'B01001_001E', geometry = TRUE)
}
