% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper_functions.R
\name{get_variables}
\alias{get_variables}
\title{Collect a frame of a specific variable from the ACS or Census}
\usage{
get_variables(var, year = 2015, dataset = "acs5")
}
\arguments{
\item{var}{The variable for collection.}

\item{year}{The year of the collection.}

\item{dataset}{The options for census bureau collection. Defaults to 'acs5'
but options include 'sf1', 'sf3', 'acs1', 'acs3' or 'acs5'}
}
\value{
A tibble of your specific variable rows.
}
\description{
Collect a frame of a specific variable from the ACS or Census
}
\examples{
get_variables('B02001')
}
\seealso{
\code{\link{tidycensus::load_variables}}
}
