% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{map_theme_soft}
\alias{map_theme_soft}
\title{Soft theme for ggplot2 maps}
\usage{
map_theme_soft(base_size = 12)
}
\arguments{
\item{base_size}{The default font size for the map.}
}
\description{
Soft theme for ggplot2 maps
}
