% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topcities.R
\docType{data}
\name{topcities}
\alias{topcities}
\title{Top US Cities by Population}
\format{dataframe}
\usage{
topcities
}
\description{
Top 300+ cities in the United States according to population.
}
\examples{
  topcities
}
\keyword{datasets}
